Make changes requested by my partner
