+++ src/modules/involvementAPI.js
@@ -1,3 +1,5 @@
+import refresh from '../popup/commentDisplay.js';
+
 export default class InvolvementAPI {
   constructor() {
     this.baseURL = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/';
@@ -17,6 +19,16 @@ export default class InvolvementAPI {
     }
   }
 
+  async getComments(id) {
+    const response = await fetch(`${this.baseURL}apps/${this.appId}/comments?item_id=${id}`);
+    if (!response.ok) {
+      throw new Error(`Error: ${response.statusText}`);
+    }
+    const comments = await response.json();
+    refresh(comments);
+    return comments;
+  }
+
   async addLike(itemId) {
     try {
       const response = await fetch(`${this.baseURL}apps/${this.appId}/likes`, {
@@ -31,6 +43,28 @@ export default class InvolvementAPI {
       if (!response.ok) {
         throw new Error(`Error: ${response.statusText}`);
       }
+    } catch (error) {
+      throw new Error(`Error:${error}`);
+    }
+  }
+
+  async addComments(id, user, usercomment) {
+    try {
+      const response = await fetch(`${this.baseURL}apps/${this.appId}/comments`, {
+        method: 'POST',
+        body: JSON.stringify({
+          item_id: id,
+          username: user,
+          comment: usercomment,
+        }),
+        headers: {
+          'Content-Type': 'application/json; charset=UTF-8',
+        },
+      });
+      if (!response.ok) {
+        throw new Error(`Error: ${response.statusText}`);
+      }
+      return response.ok;
     } catch (error) {
       throw new Error(`Error:${error}`);
     }
