+++ src/style.scss
@@ -147,7 +147,7 @@ main {
       border-radius: 0.25rem;
     }
 
-    .nameInput {
+    .name {
       font-size: 1.5rem;
       padding: 1rem 0;
       font-weight: 600;
